requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.31.0
selectolax>=0.3.14
python-dotenv>=1.0.0
cachetools>=5.3.1
pydantic>=2.0.0
